requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pyOpenSSL==23.3.0
cryptography==41.0.7
boto3==1.34.0
easyocr==1.7.0
Pillow==10.0.1